move MinAggregationResult implementation into own file
