+++ searchlib/src/vespa/searchlib/aggregation/minaggregationresult.h
@@ -7,21 +7,30 @@
 
 namespace search::aggregation {
 
+/**
+ * Aggregator that keeps the minimum value.
+ */
 class MinAggregationResult : public AggregationResult {
 public:
     using SingleResultNode = expression::SingleResultNode;
-    DECLARE_AGGREGATIONRESULT(MinAggregationResult);
-    void visitMembers(vespalib::ObjectVisitor& visitor) const override;
-    const SingleResultNode& getMin() const { return *_min; }
-    MinAggregationResult();
-    MinAggregationResult(const SingleResultNode& min);
-    ~MinAggregationResult();
 
 private:
-    const ResultNode& onGetRank() const override { return getMin(); }
+    SingleResultNode::CP _min;
+
+public:
+    DECLARE_AGGREGATIONRESULT(MinAggregationResult);
+
+    MinAggregationResult();
+    explicit MinAggregationResult(const SingleResultNode& min);
+    ~MinAggregationResult() override;
+
+    void visitMembers(vespalib::ObjectVisitor& visitor) const override;
+    [[nodiscard]] const SingleResultNode& getMin() const { return *_min; }
+
+private:
+    [[nodiscard]] const ResultNode& onGetRank() const override { return getMin(); }
     void onPrepare(const ResultNode& result) override;
     void initForUnitTest(const ResultNode& result) override;
-    SingleResultNode::CP _min;
 };
 
 } // namespace search::aggregation
